Rename the Python module to `howso_engine_rl_recipes`.
Eliminate the inconsistencies in naming throughout.
