+++ howso_engine_rl_recipes/__main__.py
@@ -4,5 +4,5 @@ from .simulation import Simulation
 
 
 if __name__ == '__main__':
-    # CLI entrypoint when run via `python -m howso_recipes_engine_rl`
+    # CLI entrypoint when run via `python -m howso_engine_rl_recipes`
     Simulation.entrypoint(sys.argv[1:])
