--- howso_recipes_engine_rl/wafer_thin_mint/__init__.py
@@ -1,6 +1,0 @@
-from gymnasium.envs.registration import register
-
-register(
-    id="WaferThinMint-v0",
-    entry_point="howso_recipes_engine_rl.wafer_thin_mint.wafer_thin_mint:WaferThinMintEnv",
-)
