+++ requirements.txt
@@ -1,0 +1,118 @@
+#
+# This file is autogenerated by pip-compile with Python 3.11
+# by the following command:
+#
+#    pip-compile --output-file=requirements.txt requirements-dev.in requirements.in
+#
+--index-url https://mkoistinen@diveplane.com:AKCp5em6f1Movb6ygY1pz3S4Ldh1XSjZ1oZzEiVcfcY5RowUuzPbGy5DWLew5ZwQhnpzFUKfe@dpbuild.jfrog.io/artifactory/api/pypi/vpypi-edge/simple
+
+amalgam-lang==2.5.0
+    # via howso-engine
+certifi==2023.7.22
+    # via howso-engine
+cffi==1.15.1
+    # via cryptography
+cloudpickle==2.2.1
+    # via gymnasium
+cryptography==41.0.3
+    # via howso-engine
+deprecation==2.1.0
+    # via howso-engine
+execnet==2.0.2
+    # via pytest-xdist
+faker==19.6.1
+    # via howso-engine
+farama-notifications==0.0.4
+    # via gymnasium
+flake8==3.9.2
+    # via -r requirements-dev.in
+gymnasium[classic-control]==0.29.1
+    # via -r requirements.in
+howso-engine==7.3.0
+    # via -r requirements.in
+howso-openapi-client==24.0.16
+    # via howso-engine
+humanize==4.8.0
+    # via howso-engine
+iniconfig==2.0.0
+    # via pytest
+joblib==1.3.2
+    # via scikit-learn
+markdown-it-py==3.0.0
+    # via rich
+mccabe==0.6.1
+    # via flake8
+mdurl==0.1.2
+    # via markdown-it-py
+mmh3==4.0.1
+    # via howso-engine
+numpy==1.23.5
+    # via
+    #   -r requirements.in
+    #   gymnasium
+    #   howso-engine
+    #   pandas
+    #   scikit-learn
+    #   scipy
+packaging==23.1
+    # via
+    #   deprecation
+    #   pytest
+pandas==1.5.0
+    # via
+    #   -r requirements.in
+    #   howso-engine
+pluggy==1.3.0
+    # via pytest
+pycodestyle==2.7.0
+    # via flake8
+pycparser==2.21
+    # via cffi
+pyflakes==2.3.1
+    # via flake8
+pygame==2.5.1
+    # via gymnasium
+pygments==2.16.1
+    # via rich
+pyjwt==2.8.0
+    # via howso-engine
+pytest==7.4.2
+    # via
+    #   -r requirements-dev.in
+    #   pytest-xdist
+pytest-xdist==3.3.1
+    # via -r requirements-dev.in
+python-dateutil==2.8.2
+    # via
+    #   faker
+    #   howso-openapi-client
+    #   pandas
+pytz==2023.3.post1
+    # via pandas
+pyyaml==6.0.1
+    # via howso-engine
+rich==13.5.2
+    # via howso-engine
+scikit-learn==1.3.0
+    # via howso-engine
+scipy==1.11.2
+    # via scikit-learn
+semantic-version==2.10.0
+    # via howso-engine
+six==1.16.0
+    # via
+    #   howso-openapi-client
+    #   python-dateutil
+threadpoolctl==3.2.0
+    # via scikit-learn
+typing-extensions==4.7.1
+    # via
+    #   gymnasium
+    #   howso-engine
+urllib3==1.26.16
+    # via
+    #   howso-engine
+    #   howso-openapi-client
+
+# The following packages are considered to be unsafe in a requirements file:
+# setuptools
