+++ tests/test_wtm.py
@@ -2,7 +2,7 @@ import logging
 import pytest
 from multiprocessing import cpu_count
 
-from howso_recipes_engine_rl.simulation import GameType, Simulation
+from howso_engine_rl_recipes.simulation import GameType, Simulation
 
 logger = logging.getLogger("howso.rl.tests")
